+++ src/base/loaders/PNTSLoaderBase.js
@@ -1,23 +1,22 @@
-// I3DM File Format
-// https://github.com/CesiumGS/3d-tiles/blob/master/specification/TileFormats/Instanced3DModel/README.md
+// PNTS File Format
+// https://github.com/CesiumGS/3d-tiles/blob/master/specification/TileFormats/PointCloud/README.md
 
-import { FeatureTable, BatchTable } from '../utilities/FeatureTable.js';
-import { arrayToString } from '../utilities/arrayToString.js';
+import { FeatureTable, BatchTable } from '../../utilities/FeatureTable.js';
+import { readMagicBytes } from '../../utilities/readMagicBytes.js';
 import { LoaderBase } from './LoaderBase.js';
-import { readMagicBytes } from '../utilities/readMagicBytes.js';
 
-export class I3DMLoaderBase extends LoaderBase {
+export class PNTSLoaderBase extends LoaderBase {
 
 	parse( buffer ) {
 
 		const dataView = new DataView( buffer );
 
-		// 32-byte header
+		// 28-byte header
 
 		// 4 bytes
 		const magic = readMagicBytes( dataView );
 
-		console.assert( magic === 'i3dm' );
+		console.assert( magic === 'pnts' );
 
 		// 4 bytes
 		const version = dataView.getUint32( 4, true );
@@ -41,11 +40,8 @@ export class I3DMLoaderBase extends LoaderBase {
 		// 4 bytes
 		const batchTableBinaryByteLength = dataView.getUint32( 24, true );
 
-		// 4 bytes
-		const gltfFormat = dataView.getUint32( 28, true );
-
 		// Feature Table
-		const featureTableStart = 32;
+		const featureTableStart = 28;
 		const featureTableBuffer = buffer.slice(
 			featureTableStart,
 			featureTableStart + featureTableJSONByteLength + featureTableBinaryByteLength,
@@ -65,53 +61,17 @@ export class I3DMLoaderBase extends LoaderBase {
 		);
 		const batchTable = new BatchTable(
 			batchTableBuffer,
-			featureTable.getData( 'INSTANCES_LENGTH' ),
+			featureTable.getData( 'BATCH_LENGTH' ) || featureTable.getData( 'POINTS_LENGTH' ),
 			0,
 			batchTableJSONByteLength,
 			batchTableBinaryByteLength,
 		);
 
-		const glbStart = batchTableStart + batchTableJSONByteLength + batchTableBinaryByteLength;
-		const bodyBytes = new Uint8Array( buffer, glbStart, byteLength - glbStart );
+		return Promise.resolve( {
 
-		let glbBytes = null;
-		let promise = null;
-		if ( gltfFormat ) {
-
-			glbBytes = bodyBytes;
-			promise = Promise.resolve();
-
-		} else {
-
-			const externalUri = this.resolveExternalURL( arrayToString( bodyBytes ) );
-			promise = fetch( externalUri, this.fetchOptions )
-				.then( res => {
-
-					if ( ! res.ok ) {
-
-						throw new Error( `I3DMLoaderBase : Failed to load file "${ externalUri }" with status ${ res.status } : ${ res.statusText }` );
-
-					}
-
-					return res.arrayBuffer();
-
-				} )
-				.then( buffer => {
-
-					glbBytes = new Uint8Array( buffer );
-
-				} );
-
-		}
-
-		return promise.then( () => {
-
-			return {
-				version,
-				featureTable,
-				batchTable,
-				glbBytes,
-			};
+			version,
+			featureTable,
+			batchTable,
 
 		} );
