+++ src/three/loaders/CMPTLoader.d.ts
@@ -1,6 +1,6 @@
-import { B3DMBaseResult } from '../../base/B3DMLoaderBase';
-import { I3DMBaseResult } from '../../base/I3DMLoaderBase';
-import { PNTSBaseResult } from '../../base/PNTSLoaderBase';
+import { B3DMBaseResult } from '../../base/loaders/B3DMLoaderBase';
+import { I3DMBaseResult } from '../../base/loaders/I3DMLoaderBase';
+import { PNTSBaseResult } from '../../base/loaders/PNTSLoaderBase';
 import { Group, LoadingManager } from 'three';
 
 export interface CMPTResult {
