+++ src/three/loaders/I3DMLoader.js
@@ -1,4 +1,4 @@
-import { I3DMLoaderBase } from '../../base/I3DMLoaderBase.js';
+import { I3DMLoaderBase } from '../../base/loaders/I3DMLoaderBase.js';
 import { DefaultLoadingManager, Matrix4, InstancedMesh, Vector3, Quaternion } from 'three';
 import { GLTFLoader } from 'three/examples/jsm/loaders/GLTFLoader.js';
